+++ src/store/repository.ts
@@ -1,15 +1,21 @@
 import type { Dirtiness } from '../commands/git/dirtiness/dirtiness';
 
 
-export type Repository = {
+/** TODO rename */
+export type LocalRepository = {
+  type: 'local';
   name: string;
-  description: string | null;
   ownerLogin: string;
-  languageName?: string; // "C++" etc
   /** GitHub project url. */
   url: string;
-  /** GitHub project url, with .git on end. */
-  // gitUrl: string;
+  localPath?: string;
+  dirty?: Dirtiness;
+};
+
+export type RemoteRepository = Omit<LocalRepository, 'type'> & {
+  type: 'remote';
+  description: string | null;
+  languageName?: string; // "C++" etc
 
   isPrivate: boolean;
   isTemplate: boolean;
@@ -21,12 +27,8 @@ export type Repository = {
   createdAt: Date;
   updatedAt: Date;
 
-  localPath?: string;
-  dirty?: Dirtiness;
   isFavorited?: boolean;
 };
 
-export type LocalRepository = {
-  dirPath: string;
-  gitUrl: string;
-};
+
+export type Repository = LocalRepository | RemoteRepository;