+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log : GitHub Repository Manager
 
+## 1.6.0
+* Added `Cloned - Others` tree for local repositories that aren't from the user or from an organization the user is part of. For now they must have a remote URL. Later I will add support for repositories without remote.
+* Added `Delete` to cloned repositories' context menu. It has a confirmation and informs if the repository is dirty, but still, be careful!
+
 ## 1.5.1 - 2022-06-20
 * Improved HEAD detection for repository cloning. It nall shall work for all the system languages. [#42](https://github.com/SrBrahma/GitHub-Repository-Manager/issues/42). Thanks, [lo-ca](https://github.com/lo-ca)!